+++ Sort.h
@@ -18,9 +18,9 @@ using namespace std;
 #define flash 10
 
 // Selection sort
-int indexMax(int *arr, int size, long int &comparisions);
+int indexMax(int *arr, int size, long long int &comparisions);
 void swap(int &a, int &b);
-void selectionSort(int *&arr, int size, long int &comparisions);
+void selectionSort(int *&arr, int size, long long int &comparisions);
 
 void insertionSort(int *arr, int size);
 void bubbleSort(int *arr, int size);
@@ -32,4 +32,4 @@ void shakerSort(int *arr, int size);
 void shellSort(int *arr, int size);
 void countingSort(int *arr, int size);
 void flashSort(int arr, int size);
-void SortData(int type_sort, clock_t &start, clock_t &end, long int &comparisions, int *&arr, int size);
+void SortData(int type_sort, clock_t &start, clock_t &end, long long int &comparisions, int *&arr, int size);
